requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
emoji==2.2.0
emoji
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
tensorflow==2.12.0
textblob==0.17.1
Pillow==9.5.0
aiohttp==3.8.4
